+++ service/hub/src/main/java/com/sparta/hub/application/service/HubService.java
@@ -1,9 +1,9 @@
 package com.sparta.hub.application.service;
 
-import com.sparta.hub.application.dto.HubCreateRequest;
-import com.sparta.hub.application.dto.HubResponse;
-import com.sparta.hub.application.dto.HubSearchCond;
-import com.sparta.hub.application.dto.HubUpdateRequest;
+import com.sparta.hub.application.dto.hub.HubCreateRequest;
+import com.sparta.hub.application.dto.hub.HubResponse;
+import com.sparta.hub.application.dto.hub.HubSearchCond;
+import com.sparta.hub.application.dto.hub.HubUpdateRequest;
 import com.sparta.hub.application.mapper.HubMapper;
 import com.sparta.hub.domain.Hub;
 import com.sparta.hub.exception.AlreadyDeletedException;
